Changing Count Data Type

+++ java/engine/org/apache/derby/impl/sql/compile/CountAggregateDefinition.java
@@ -61,7 +61,7 @@ public class CountAggregateDefinition
 		/*
 		** COUNT never returns NULL
 		*/
-		return DataTypeDescriptor.getBuiltInDataTypeDescriptor(java.sql.Types.INTEGER, false);
+		return DataTypeDescriptor.getBuiltInDataTypeDescriptor(java.sql.Types.BIGINT, false);
 	}
 
 }
